+++ src/client/widgets/scripts/script-vm.js
@@ -77,6 +77,7 @@ class ScriptVm extends Vm {
             if (extraOptions.send === false) options.send = false
             if (extraOptions.sync === false) options.sync = false
             if (extraOptions.script === false) options.script = false
+            if (extraOptions.external === true) options.fromExternal = true
 
             var widgets
             if (id.includes('*')) {